requirements.txt
requests==2.31.0
python-dotenv==1.0.0
discord.py==2.3.2
pyTelegramBotAPI==4.14.0
